feat(calibration): generalize multi-gauge infrastructure for HYPE, FUSE, and SUMMA
Move MultiGaugeMetrics base class and gauge-segment mapping logic to a
shared optimization/multi_gauge/ package so all three models can use
distributed multi-gauge calibration without cross-model imports.

- Move MultiGaugeMetrics to optimization/multi_gauge/metrics.py
- Extract ensure_gauge_mapping() to optimization/multi_gauge/gauge_mapping.py
- Add _calculate_multi_gauge_metrics() to SUMMAWorker (uses mizuRoute)
- Add elev_m to climate statistics CSV from catchment shapefile
- FUSE/HYPE modules re-export for backward compatibility
