+++ src/symfluence/models/summa/calibration/worker.py
@@ -323,6 +323,24 @@ class SUMMAWorker(BaseWorker):
             Dictionary of metric names to values
         """
         try:
+            multi_gauge_enabled = config.get('MULTI_GAUGE_CALIBRATION', False)
+            if multi_gauge_enabled and self.needs_routing(config):
+                mizuroute_dir = kwargs.get('mizuroute_dir')
+                if not mizuroute_dir:
+                    sim_dir = kwargs.get('sim_dir', output_dir)
+                    mizuroute_dir = sim_dir / 'mizuRoute'
+                mizuroute_output = self._find_mizuroute_output(Path(mizuroute_dir))
+                if mizuroute_output is not None:
+                    data_dir = Path(config.get('SYMFLUENCE_DATA_DIR', '.'))
+                    domain_name = config.get('DOMAIN_NAME', '')
+                    project_dir = data_dir / f'domain_{domain_name}'
+                    return self._calculate_multi_gauge_metrics(
+                        config=config,
+                        mizuroute_output_path=mizuroute_output,
+                        project_dir=project_dir,
+                        settings_dir=kwargs.get('settings_dir'),
+                    )
+
             # Import existing function
             from symfluence.optimization.workers.summa import _calculate_metrics_with_target
 
@@ -350,6 +368,184 @@ class SUMMAWorker(BaseWorker):
         except Exception as e:  # noqa: BLE001 — calibration resilience
             self.logger.error(f"Error calculating SUMMA metrics: {e}")
             return {'kge': self.penalty_score}
+
+    # =========================================================================
+    # Multi-Gauge Metrics (delegates to shared MultiGaugeMetrics)
+    # =========================================================================
+
+    def _calculate_multi_gauge_metrics(
+        self,
+        config: Dict[str, Any],
+        mizuroute_output_path: Path,
+        project_dir: Path,
+        **kwargs
+    ) -> Dict[str, Any]:
+        """Calculate performance metrics across multiple stream gauges.
+
+        Uses the shared :class:`MultiGaugeMetrics` to read per-segment
+        routed discharge from mizuRoute output and compare against
+        LaMAH-Ice gauge observations.
+        """
+        try:
+            from symfluence.optimization.multi_gauge.metrics import MultiGaugeMetrics
+
+            gauge_mapping_path = config.get('GAUGE_SEGMENT_MAPPING')
+            obs_dir = config.get('MULTI_GAUGE_OBS_DIR')
+            gauge_ids = config.get('MULTI_GAUGE_IDS')
+            exclude_ids = config.get('MULTI_GAUGE_EXCLUDE_IDS', [])
+            aggregation = config.get('MULTI_GAUGE_AGGREGATION', 'mean')
+            min_gauges = config.get('MULTI_GAUGE_MIN_GAUGES', 5)
+
+            # Auto-generate gauge mapping if not configured
+            if not gauge_mapping_path:
+                gauge_mapping_path = self._ensure_gauge_mapping(config, project_dir)
+            if not gauge_mapping_path:
+                self.logger.error("GAUGE_SEGMENT_MAPPING not configured for multi-gauge calibration")
+                return {'kge': self.penalty_score}
+            if not obs_dir:
+                self.logger.error("MULTI_GAUGE_OBS_DIR not configured for multi-gauge calibration")
+                return {'kge': self.penalty_score}
+
+            gauge_mapping_path = Path(gauge_mapping_path)
+            obs_dir = Path(obs_dir)
+
+            if 'D_gauges' in obs_dir.parts and not obs_dir.exists():
+                try:
+                    from symfluence.data.observation.handlers.lamah_ice import (
+                        ensure_lamah_ice_streamflow,
+                    )
+                    lamah_root = obs_dir
+                    while lamah_root.name != 'D_gauges' and lamah_root.parent != lamah_root:
+                        lamah_root = lamah_root.parent
+                    lamah_root = lamah_root.parent
+                    ensure_lamah_ice_streamflow(lamah_root, self.logger)
+                except Exception as exc:  # noqa: BLE001
+                    self.logger.warning(f"LaMAH-Ice auto-download skipped: {exc}")
+
+            topology_path = self._find_topology_path(kwargs.get('settings_dir'))
+            start_date, end_date = self._parse_calibration_period(config)
+
+            multi_gauge = MultiGaugeMetrics(
+                gauge_segment_mapping_path=gauge_mapping_path,
+                obs_data_dir=obs_dir,
+                logger=self.logger,
+            )
+
+            if gauge_ids is None and exclude_ids:
+                all_gauge_ids = multi_gauge.gauge_mapping['id'].tolist()
+                gauge_ids = [gid for gid in all_gauge_ids if gid not in exclude_ids]
+
+            filter_config = {}
+            for key, cfg_key in [('max_distance', 'MULTI_GAUGE_MAX_DISTANCE'),
+                                  ('min_obs_cv', 'MULTI_GAUGE_MIN_OBS_CV'),
+                                  ('min_specific_q', 'MULTI_GAUGE_MIN_SPECIFIC_Q')]:
+                val = config.get(cfg_key)
+                if val is not None:
+                    filter_config[key] = float(val)
+
+            min_overlap = int(config.get('MULTI_GAUGE_MIN_OVERLAP_DAYS', 10))
+            kge_floor = config.get('MULTI_GAUGE_KGE_FLOOR')
+            if kge_floor is not None:
+                kge_floor = float(kge_floor)
+
+            results = multi_gauge.calculate_multi_gauge_metrics(
+                mizuroute_output_path=mizuroute_output_path,
+                gauge_ids=gauge_ids,
+                start_date=start_date,
+                end_date=end_date,
+                topology_path=topology_path,
+                min_gauges=min_gauges,
+                aggregation=aggregation,
+                filter_config=filter_config if filter_config else None,
+                min_overlap_days=min_overlap,
+                kge_floor=kge_floor,
+            )
+
+            self.logger.debug(
+                f"Multi-gauge calibration: KGE={results['kge']:.4f} "
+                f"({results['n_valid_gauges']}/{results['n_total_gauges']} valid gauges)"
+            )
+
+            return {
+                'kge': results['kge'],
+                'kge_std': results.get('kge_std', 0.0),
+                'kge_min': results.get('kge_min', results['kge']),
+                'kge_max': results.get('kge_max', results['kge']),
+                'n_gauges': results['n_valid_gauges'],
+                'multi_gauge_details': results.get('per_gauge', {}),
+            }
+
+        except FileNotFoundError as e:
+            self.logger.error(f"Multi-gauge file not found: {e}")
+            return {'kge': self.penalty_score}
+        except Exception as e:  # noqa: BLE001 — calibration resilience
+            self.logger.error(f"Error in multi-gauge metrics calculation: {e}")
+            return {'kge': self.penalty_score}
+
+    def _ensure_gauge_mapping(
+        self,
+        config: Dict[str, Any],
+        project_dir: Path,
+    ) -> Optional[Path]:
+        """Auto-generate gauge-segment mapping if it doesn't exist."""
+        lamah_path = config.get('LAMAH_ICE_PATH')
+        if not lamah_path:
+            data_dir = config.get('SYMFLUENCE_DATA_DIR')
+            if data_dir:
+                lamah_path = Path(data_dir) / 'lamah_ice'
+        if not lamah_path:
+            return None
+        lamah_path = Path(lamah_path)
+        if lamah_path.name == 'default':
+            data_dir = config.get('SYMFLUENCE_DATA_DIR')
+            if data_dir:
+                lamah_path = Path(data_dir) / 'lamah_ice'
+
+        domain_name = config.get('DOMAIN_NAME', '')
+        try:
+            from symfluence.optimization.multi_gauge.gauge_mapping import ensure_gauge_mapping
+            return ensure_gauge_mapping(
+                project_dir, lamah_path, domain_name,
+                output_subdir='mizuRoute',
+                output_filename='gauge_segment_mapping.csv',
+                logger=self.logger,
+            )
+        except Exception as e:  # noqa: BLE001
+            self.logger.warning(f"Could not auto-generate gauge mapping: {e}")
+            return None
+
+    def _find_mizuroute_output(self, mizuroute_dir: Path) -> Optional[Path]:
+        """Find the mizuRoute output NetCDF file."""
+        if not mizuroute_dir.exists():
+            return None
+        for pattern in ('*.nc', '*output*.nc'):
+            files = sorted(mizuroute_dir.glob(pattern))
+            if files:
+                return files[-1]
+        return None
+
+    def _find_topology_path(self, settings_dir: Any) -> Optional[Path]:
+        """Find mizuRoute topology file."""
+        if not settings_dir:
+            return None
+        settings_dir = Path(settings_dir)
+        mizu_settings = settings_dir / 'mizuRoute'
+        if mizu_settings.exists():
+            for name in ('topology.nc', 'network_topology.nc'):
+                path = mizu_settings / name
+                if path.exists():
+                    return path
+        return None
+
+    def _parse_calibration_period(self, config: Dict[str, Any]) -> tuple:
+        """Parse calibration period from config."""
+        cal_period = config.get('CALIBRATION_PERIOD', '')
+        if isinstance(cal_period, str) and ',' in cal_period:
+            parts = [p.strip() for p in cal_period.split(',')]
+            return parts[0], parts[1]
+        start = config.get('EXPERIMENT_TIME_START', '')
+        end = config.get('EXPERIMENT_TIME_END', '')
+        return str(start).strip(), str(end).strip()
 
     def _calculate_metrics_direct(
         self,
