+++ src/symfluence/optimization/multi_gauge/__init__.py
@@ -1,0 +1,4 @@
+# SPDX-License-Identifier: GPL-3.0-or-later
+# Copyright (C) 2024-2026 SYMFLUENCE Team <dev@symfluence.org>
+
+"""Shared multi-gauge calibration utilities."""
